Fix for creating IDs for roles on 64bit unix systems
Fixed a bug causing the admin role not get initialized properly after new
installations on 64bit unix systems.
Reason is the way the database ID is generated for roles: the crc32()
functions is used on the role's key.
In 64bit unix environments this function may return values bigger than
signed 32bit, causing an out of range error on the database.
Added a workaround so that crc32() function only returns 32bit values.